fix(ai): split below-quota peace part2 tests for 1000-line gate
Move tests 12–18 into part3 so repo.dart_file_non_comment_line_size
passes on PR #2570 without changing test coverage.
