+++ packages/colonizethis_ai/test/diplomacy_planner_below_quota_peace_part2_test.dart
@@ -650,530 +650,4 @@ void main() {
       );
     },
   );
-
-  test(
-    'shouldSkipBelowQuotaGpOnlyBlockerPeacePass false below observer quota at 8 OW',
-    () {
-      final game = Game(
-        id: 'g-skip-gp-only-peace-lead',
-        worldState: WorldState(
-          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 50),
-          oldWorld: RegionData(
-            provinces: [
-              for (var i = 1; i <= 8; i++)
-                Province(
-                  id: 'oldWorld|gp5_$i',
-                  regionId: 'oldWorld',
-                  ownerId: 'gp5',
-                ),
-              for (var i = 1; i <= 10; i++)
-                Province(
-                  id: 'oldWorld|gp6_$i',
-                  regionId: 'oldWorld',
-                  ownerId: 'gp6',
-                ),
-              const Province(
-                id: 'oldWorld|frontier',
-                regionId: 'oldWorld',
-                ownerId: 'gp6',
-              ),
-            ],
-          ),
-          newWorld: const RegionData(),
-        ),
-        players: const [
-          Player(id: 'gp5', displayName: 'P5', isHuman: false),
-          Player(id: 'gp6', displayName: 'P6', isHuman: false),
-        ],
-        diplomacyRelations: [
-          const DiplomacyRelation(
-            factionId1: 'gp5',
-            factionId2: 'gp6',
-            state: RelationState.atWar,
-            score: 30,
-          ),
-        ],
-      );
-      const snapshot = AIWorldSnapshot(
-        playerId: 'gp5',
-        threats: ThreatSummary(atWarWith: ['gp6']),
-        opportunities: OpportunitySummary(),
-        conquest: ConquestSummary(
-          oldWorldProvincesOwned: 8,
-          invadableProvinceIdsSorted: ['oldWorld|frontier'],
-        ),
-        economy: EconomySummary(),
-        relations: {},
-      );
-      expect(
-        shouldSkipBelowQuotaGpOnlyBlockerPeacePass(
-          game: game,
-          snapshot: snapshot,
-        ),
-        isFalse,
-      );
-    },
-  );
-
-  test(
-    'stalledZeroRegimentGpPeaceTargets includes all GP wars when stalled',
-    () {
-      final game = Game(
-        id: 'g-zero-reg-gp-peace',
-        worldState: WorldState(
-          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 80),
-          oldWorld: RegionData(
-            provinces: [
-              for (var i = 1; i <= 8; i++)
-                Province(
-                  id: 'oldWorld|gp3_$i',
-                  regionId: 'oldWorld',
-                  ownerId: 'gp3',
-                ),
-              for (var i = 1; i <= 12; i++)
-                Province(
-                  id: 'oldWorld|gp4_$i',
-                  regionId: 'oldWorld',
-                  ownerId: 'gp4',
-                ),
-            ],
-          ),
-          newWorld: const RegionData(),
-        ),
-        players: const [
-          Player(id: 'gp3', displayName: 'P3', isHuman: false),
-          Player(id: 'gp4', displayName: 'P4', isHuman: false),
-        ],
-        diplomacyRelations: [
-          const DiplomacyRelation(
-            factionId1: 'gp3',
-            factionId2: 'gp4',
-            state: RelationState.atWar,
-            score: 30,
-          ),
-        ],
-      );
-      const snapshot = AIWorldSnapshot(
-        playerId: 'gp3',
-        threats: ThreatSummary(atWarWith: ['gp4', 'minor1']),
-        opportunities: OpportunitySummary(),
-        conquest: ConquestSummary(
-          oldWorldProvincesOwned: 8,
-          invadableProvinceIdsSorted: ['oldWorld|inv1'],
-        ),
-        economy: EconomySummary(),
-        relations: {},
-      );
-      expect(
-        stalledZeroRegimentGpPeaceTargets(game: game, snapshot: snapshot),
-        ['gp4'],
-      );
-    },
-  );
-
-  test(
-    'plateauMutualInvadableBlockerPeaceTargets peace mutual below-quota blockers',
-    () {
-      final game = Game(
-        id: 'g-plateau-mutual-blocker',
-        worldState: WorldState(
-          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 60),
-          oldWorld: RegionData(
-            provinces: [
-              for (var i = 1; i <= 8; i++)
-                Province(
-                  id: 'oldWorld|gp5_$i',
-                  regionId: 'oldWorld',
-                  ownerId: 'gp5',
-                ),
-              for (var i = 1; i <= 8; i++)
-                Province(
-                  id: 'oldWorld|gp6_$i',
-                  regionId: 'oldWorld',
-                  ownerId: 'gp6',
-                ),
-              const Province(
-                id: 'oldWorld|inv1',
-                regionId: 'oldWorld',
-                ownerId: 'gp6',
-              ),
-              const Province(
-                id: 'oldWorld|inv2',
-                regionId: 'oldWorld',
-                ownerId: 'minor1',
-              ),
-            ],
-          ),
-          newWorld: const RegionData(),
-        ),
-        players: const [
-          Player(id: 'gp5', displayName: 'P5', isHuman: false),
-          Player(id: 'gp6', displayName: 'P6', isHuman: false),
-        ],
-        minorNations: const [MinorNation(id: 'minor1', displayName: 'M1')],
-        diplomacyRelations: [
-          const DiplomacyRelation(
-            factionId1: 'gp5',
-            factionId2: 'gp6',
-            state: RelationState.atWar,
-            score: 30,
-          ),
-        ],
-      );
-      const snapshot = AIWorldSnapshot(
-        playerId: 'gp5',
-        threats: ThreatSummary(atWarWith: ['gp6']),
-        opportunities: OpportunitySummary(),
-        conquest: ConquestSummary(
-          oldWorldProvincesOwned: 8,
-          invadableProvinceIdsSorted: ['oldWorld|inv1', 'oldWorld|inv2'],
-        ),
-        economy: EconomySummary(),
-        relations: {},
-      );
-      expect(
-        plateauMutualInvadableBlockerPeaceTargets(game: game, snapshot: snapshot),
-        isEmpty,
-        reason: 'below-quota mutual blocker peace stalls observer conquest',
-      );
-    },
-  );
-
-  test(
-    'weakHoldingsInvadableBlockerPeaceTargets peace frontier GP when outmatched',
-    () {
-      final game = Game(
-        id: 'g-weak-blocker-peace',
-        worldState: WorldState(
-          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 60),
-          oldWorld: RegionData(
-            provinces: [
-              for (var i = 1; i <= 4; i++)
-                Province(
-                  id: 'oldWorld|gp3_$i',
-                  regionId: 'oldWorld',
-                  ownerId: 'gp3',
-                ),
-              for (var i = 1; i <= 11; i++)
-                Province(
-                  id: 'oldWorld|gp4_$i',
-                  regionId: 'oldWorld',
-                  ownerId: 'gp4',
-                ),
-              const Province(
-                id: 'oldWorld|inv1',
-                regionId: 'oldWorld',
-                ownerId: 'gp4',
-              ),
-              const Province(
-                id: 'oldWorld|inv2',
-                regionId: 'oldWorld',
-                ownerId: 'minor1',
-              ),
-            ],
-          ),
-          newWorld: const RegionData(),
-        ),
-        players: const [
-          Player(id: 'gp3', displayName: 'P3', isHuman: false),
-          Player(id: 'gp4', displayName: 'P4', isHuman: false),
-        ],
-        minorNations: const [MinorNation(id: 'minor1', displayName: 'M1')],
-        diplomacyRelations: [
-          const DiplomacyRelation(
-            factionId1: 'gp3',
-            factionId2: 'gp4',
-            state: RelationState.atWar,
-            score: 30,
-          ),
-        ],
-      );
-      const snapshot = AIWorldSnapshot(
-        playerId: 'gp3',
-        threats: ThreatSummary(atWarWith: ['gp4']),
-        opportunities: OpportunitySummary(),
-        conquest: ConquestSummary(
-          oldWorldProvincesOwned: 4,
-          invadableProvinceIdsSorted: ['oldWorld|inv1', 'oldWorld|inv2'],
-        ),
-        economy: EconomySummary(),
-        relations: {},
-      );
-      expect(
-        weakHoldingsInvadableBlockerPeaceTargets(game: game, snapshot: snapshot),
-        ['gp4'],
-      );
-    },
-  );
-
-  test(
-    'criticalWeakUninvadedMinorDeclareTarget picks uninvaded minor owner',
-    () {
-      final game = Game(
-        id: 'g-critical-minor-declare',
-        worldState: WorldState(
-          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 40),
-          oldWorld: RegionData(
-            provinces: [
-              for (var i = 1; i <= 5; i++)
-                Province(
-                  id: 'oldWorld|gp3_$i',
-                  regionId: 'oldWorld',
-                  ownerId: 'gp3',
-                ),
-              const Province(
-                id: 'oldWorld|m1',
-                regionId: 'oldWorld',
-                ownerId: 'minor1',
-              ),
-              const Province(
-                id: 'oldWorld|m2',
-                regionId: 'oldWorld',
-                ownerId: 'minor2',
-              ),
-            ],
-          ),
-          newWorld: const RegionData(),
-        ),
-        players: const [Player(id: 'gp3', displayName: 'P3', isHuman: false)],
-        minorNations: const [
-          MinorNation(id: 'minor1', displayName: 'M1'),
-          MinorNation(id: 'minor2', displayName: 'M2'),
-        ],
-        diplomacyRelations: [
-          const DiplomacyRelation(
-            factionId1: 'gp3',
-            factionId2: 'minor1',
-            state: RelationState.atWar,
-            score: 30,
-          ),
-        ],
-      );
-      const snapshot = AIWorldSnapshot(
-        playerId: 'gp3',
-        threats: ThreatSummary(atWarWith: ['minor1']),
-        opportunities: OpportunitySummary(),
-        conquest: ConquestSummary(
-          oldWorldProvincesOwned: 6,
-          invadableProvinceIdsSorted: ['oldWorld|m1', 'oldWorld|m2'],
-        ),
-        economy: EconomySummary(),
-        relations: {},
-      );
-      expect(
-        criticalWeakUninvadedMinorDeclareTarget(game: game, snapshot: snapshot),
-        'minor2',
-      );
-    },
-  );
-
-  test(
-    'criticalWeakUninvadedMinorDeclareTarget fires at 8 OW with sole GP war',
-    () {
-      final game = Game(
-        id: 'g-critical-minor-with-gp-war',
-        worldState: WorldState(
-          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 40),
-          oldWorld: RegionData(
-            provinces: [
-              for (var i = 1; i <= 8; i++)
-                Province(
-                  id: 'oldWorld|gp5_$i',
-                  regionId: 'oldWorld',
-                  ownerId: 'gp5',
-                ),
-              for (var i = 1; i <= 8; i++)
-                Province(
-                  id: 'oldWorld|gp6_$i',
-                  regionId: 'oldWorld',
-                  ownerId: 'gp6',
-                ),
-              const Province(
-                id: 'oldWorld|m1',
-                regionId: 'oldWorld',
-                ownerId: 'minor1',
-              ),
-            ],
-          ),
-          newWorld: const RegionData(),
-        ),
-        players: const [
-          Player(id: 'gp5', displayName: 'P5', isHuman: false),
-          Player(id: 'gp6', displayName: 'P6', isHuman: false),
-        ],
-        minorNations: const [MinorNation(id: 'minor1', displayName: 'M1')],
-        diplomacyRelations: [
-          const DiplomacyRelation(
-            factionId1: 'gp5',
-            factionId2: 'gp6',
-            state: RelationState.atWar,
-            score: 30,
-          ),
-        ],
-      );
-      const snapshot = AIWorldSnapshot(
-        playerId: 'gp5',
-        threats: ThreatSummary(atWarWith: ['gp6']),
-        opportunities: OpportunitySummary(),
-        conquest: ConquestSummary(
-          oldWorldProvincesOwned: 8,
-          invadableProvinceIdsSorted: ['oldWorld|m1'],
-        ),
-        economy: EconomySummary(),
-        relations: {},
-      );
-      expect(
-        criticalWeakUninvadedMinorDeclareTarget(game: game, snapshot: snapshot),
-        'minor1',
-      );
-    },
-  );
-
-  test(
-    'runDiplomacyPlannerWithResult forces peace when candidates are empty',
-    () {
-      final game = Game(
-        id: 'g-empty-candidates-peace',
-        worldState: WorldState(
-          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 60),
-          oldWorld: RegionData(
-            provinces: [
-              for (var i = 1; i <= 5; i++)
-                Province(
-                  id: 'oldWorld|gp3_$i',
-                  regionId: 'oldWorld',
-                  ownerId: 'gp3',
-                ),
-              for (var i = 1; i <= 10; i++)
-                Province(
-                  id: 'oldWorld|gp4_$i',
-                  regionId: 'oldWorld',
-                  ownerId: 'gp4',
-                ),
-            ],
-          ),
-          newWorld: const RegionData(),
-        ),
-        players: const [
-          Player(id: 'gp3', displayName: 'P3', isHuman: false),
-          Player(id: 'gp4', displayName: 'P4', isHuman: false),
-        ],
-        diplomacyRelations: [
-          const DiplomacyRelation(
-            factionId1: 'gp3',
-            factionId2: 'gp4',
-            state: RelationState.atWar,
-            score: 30,
-          ),
-        ],
-        aiControlByGpId: const {'gp3': true},
-      );
-      const topology = MapTopology(nodes: [], edges: []);
-      final ctx = buildTestPlannerContext(
-        game: game,
-        topology: topology,
-        nationId: 'gp3',
-        primaryGoal: StrategicGoal.trade,
-        suggestionAPI: const FakeOrderSuggestionAPIForDomainPlannerTests(
-          work: [],
-          build: [],
-          move: [],
-          research: [],
-          navalMove: [],
-          navalMission: [],
-          diplomatic: [],
-        ),
-      );
-      const snapshot = AIWorldSnapshot(
-        playerId: 'gp3',
-        threats: ThreatSummary(atWarWith: ['gp4']),
-        opportunities: OpportunitySummary(),
-        conquest: ConquestSummary(
-          oldWorldProvincesOwned: 5,
-          invadableProvinceIdsSorted: ['oldWorld|gp4_10'],
-        ),
-        economy: EconomySummary(),
-        relations: {},
-      );
-      final result = runDiplomacyPlannerWithResult(
-        ctx: ctx,
-        snapshot: snapshot,
-        pass: DiplomacyPlannerPass.nonDeclareWarOnly,
-      );
-      final peaceOrders =
-          result.orders.diplomaticOrdersByPlayerId['gp3'] ?? const [];
-      expect(
-        peaceOrders.any(
-          (o) =>
-              o.type == DiplomaticOrderType.offerPeace &&
-              o.targetFactionId == 'gp4',
-        ),
-        isFalse,
-        reason: 'below-quota GP-only frontier keeps war on invadable blocker',
-      );
-    },
-  );
-
-  test(
-    'plateauOwMinorDeclareTarget allows mutual plateau war on mixed frontier',
-    () {
-      final game = Game(
-        id: 'g-plateau-minor-mixed-frontier',
-        worldState: WorldState(
-          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 80),
-          oldWorld: RegionData(
-            provinces: [
-              for (var i = 1; i <= 8; i++)
-                Province(
-                  id: 'oldWorld|gp5_$i',
-                  regionId: 'oldWorld',
-                  ownerId: 'gp5',
-                ),
-              for (var i = 1; i <= 9; i++)
-                Province(
-                  id: 'oldWorld|gp6_$i',
-                  regionId: 'oldWorld',
-                  ownerId: 'gp6',
-                ),
-              const Province(
-                id: 'oldWorld|minor2',
-                regionId: 'oldWorld',
-                ownerId: 'minor2',
-              ),
-            ],
-          ),
-          newWorld: const RegionData(),
-        ),
-        players: const [
-          Player(id: 'gp5', displayName: 'P5', isHuman: false),
-          Player(id: 'gp6', displayName: 'P6', isHuman: false),
-        ],
-        minorNations: const [MinorNation(id: 'minor2', displayName: 'M2')],
-        diplomacyRelations: [
-          const DiplomacyRelation(
-            factionId1: 'gp5',
-            factionId2: 'gp6',
-            state: RelationState.atWar,
-            score: 30,
-          ),
-        ],
-      );
-      const snapshot = AIWorldSnapshot(
-        playerId: 'gp5',
-        threats: ThreatSummary(atWarWith: ['gp6']),
-        opportunities: OpportunitySummary(),
-        conquest: ConquestSummary(
-          oldWorldProvincesOwned: 8,
-          invadableProvinceIdsSorted: ['oldWorld|minor2'],
-          adjacentOwnerFactionIdsSorted: ['minor2'],
-        ),
-        economy: EconomySummary(),
-        relations: {},
-      );
-      expect(
-        plateauOwMinorDeclareTarget(game: game, snapshot: snapshot),
-        'minor2',
-      );
-    },
-  );
 }
